--- a/Cargo.toml
+++ b/Cargo.toml
@@ -20,6 +20,7 @@ optional = true
 reqwest = "0.8.5"
 criterion = "0.2"
 tempfile = "3.0.2"
+rand = "0.5.0"
 
 [features]
 default = ["mmap"]