--- a/benches/phrase.rs
+++ b/benches/phrase.rs
@@ -148,6 +148,27 @@ pub fn benchmark(c: &mut Criterion) {
         });
     }));
 
+    // data is shadowed here for ease of copying and pasting, but this is a new clone
+    // (again, same data, new reference, because it's an Rc)
+    let data = shared_data.clone();
+    to_bench.push(Fun::new("range_fst_range", move |b: &mut Bencher, _i| {
+        let mut rng = thread_rng();
+
+        b.iter(|| {
+            let word_ids = rng.choose(&data.phrases).unwrap();
+            let fullword_ids = &word_ids[..word_ids.len()];
+            let last_id = &word_ids[word_ids.len()-1];
+            let last_id_min = 0.max(last_id - 50);
+            let last_id_max = last_id + 50;
+            let mut query_words = fullword_ids.iter()
+                .map(|w| QueryWord::Full{ id: *w, edit_distance: 0})
+                .collect::<Vec<QueryWord>>();
+            query_words.push(QueryWord::Prefix{ id_range: (last_id_min, last_id_max) });
+            let query_phrase = QueryPhrase::new(&query_words).unwrap();
+            let _result = data.phrase_set.range(query_phrase).unwrap();
+        });
+    }));
+
     // run the accumulated list of benchmarks
     c.bench_functions("phrase", to_bench, ());
 }