This PR adds some capabilities to running phrase benchmarks on real-world data. It also will add a few new benchmarks to compare the custom solution to a straightforward use of `range` and `StartsWith` functions available from the `fst` library.

New things:

- a script for downloading phrase benchmark data and running phrase benchmarks (can be extended to support others)
- now possible to specify the source for phrase benchmark data using the env variable `PHRASE_BENCH`