@boblannon 2018-05-25
Some surprising findings so far.

| search | average time |
|--- |--- |
| `contains` | 4.3358us |
| `contains_prefix` (all full words) | 3.0256us |
| `contains_prefix` (terminal prefix) | 3.2497us |
| `fst::Set::range` (terminal prefix) | 4.7543us |

Observations:
- The straightforward `contains` query (which uses `fst::Set::contains`) is slower than the `contains_prefix` query.  surprising....
- We're doing better than the built-in range, but not as much as I thought. Maybe your experimental branches will widen that gap, @apendleton. 


<details>
<summary>Bench output</summary>

```
phrase/exact_contains   time:   [3.9365 us 4.3358 us 4.7875 us]
                        change: [+340.55% +418.44% +513.10%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe

phrase/exact_contains_prefix
                        time:   [3.0023 us 3.0256 us 3.0578 us]
                        change: [+338.42% +350.55% +362.21%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  3 (3.00%) high severe
phrase/range_contains_prefix
                        time:   [3.2125 us 3.2497 us 3.2894 us]
                        change: [+285.59% +290.76% +295.85%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 8 outliers among 100 measurements (8.00%)
  6 (6.00%) high mild
  2 (2.00%) high severe
phrase/range_fst_range  time:   [4.6648 us 4.7543 us 4.8766 us]
                        change: [+145.50% +153.99% +163.62%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 15 outliers among 100 measurements (15.00%)
  4 (4.00%) high mild
  11 (11.00%) high severe
```

</details>