@boblannon 2018-05-29
updated stats: after c0fb8bdf all methods' performance improved.

| search | avg time (old) | avg time (new) |
|--- |--- |--- |
| `contains` | 4.3358us | 667.47ns |
| `contains_prefix` (all full words) | 3.0256us | 665.13ns |
| `contains_prefix` (terminal prefix) | 3.2497us | 850.13 ns |
| `fst::Set::range` (terminal prefix) | 4.7543us | 1.8865 us | 

okay, mystery solved, sortof. looking at the flame graph, it was clear that calls to `QueryPhrase::full_word_key` were taking up a substantial amount of CPU.

![2018-05-29 at 3 32 pm](https://user-images.githubusercontent.com/1092694/40680920-a421966a-6355-11e8-9eaf-9bbfeed7c02f.png)

This change made a big difference, and I'm going to write it out so that I can try to understand exactly why.

```diff
diff --git a/src/phrase/query.rs b/src/phrase/query.rs
index db2e874..a2e747d 100644
--- a/src/phrase/query.rs
+++ b/src/phrase/query.rs
@@ -93,16 +93,17 @@ impl<'a> QueryPhrase<'a> {

     /// Generate a key from the ids of the full words in this phrase
     pub fn full_word_key(&self) -> Vec<u8> {
-        let mut word_ids: Vec<u32> = vec![];
+        let mut full_word_key: Vec<u8> = vec![];
         for word in self.words {
             match word {
                 QueryWord::Full{ ref id, .. } => {
-                    word_ids.push(*id);
+                    let three_bytes = util::three_byte_encode(*id);
+                    full_word_key.extend(three_bytes);
                 },
                 _ => (),
             }
         }
-        util::word_ids_to_key(&word_ids)
+        full_word_key
     }
```

Before this change, `full_word_key` did the following:

1. initialize `word_ids`, an empty `Vec<32>`
2. for each `QueryWord::Full` in the phrase, push the `id` value into `word_ids`
3. send `word_ids` (as a slice) to `util::word_ids_to_key`, and return the results.

fwiw, that util function is pretty straightforward:

```rust
pub fn word_ids_to_key(phrase: &[u32]) -> Vec<u8> {
    phrase.into_iter()
          .flat_map(|word| three_byte_encode(*word))
          .collect()
}
```

After the change, it works slightly differently:

1. initialize `full_word_key`, an empty `Vec<u8>`
2. for each `QueryWord::Full` in the phrase, extend that vector using the result of `three_byte_encode(word.id)`
3. return `full_word_key`

My guess as to why the new way is faster: the `flat_map` call was slowing things down because each value was being dereferenced. This would explain why the `exact_contains` benchmark was impacted more than `range_contains_prefix`. The `range_contains_prefix` benchmark calls `contains_prefix`, and we treat the final word in the query as a `QueryWord::Prefix`, so there are less `QueryWord::Full` values to resolve.

It doesn't explain, though, why (in the old code) the `exact_contains_prefix` benchmark wasn't impacted just as much as the `exact_contains` benchmark. It also calls `contains_prefix`, but all of the query words are `QueryWord::Full`, so we shouldn't expect its performance to differ from `exact_contains` results.