@boblannon 2018-05-29
...actually, i guess it's not so much the dereferencing, but just the fact that there's an extra loop for each call to `util::word_ids_to_key`.